Move words list adapter class to words view file and make it private
